requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0
pydantic>=2.0.0
loguru>=0.7.0
psutil>=5.9.0
